updating documentation
